requirements.txt
google-generativeai
langchain
streamlit
python-dotenv
PyPDF2
setuptools
langchain-google-genai
langchain_classic
langchain_core

-e .